Changed how flags and regs are represented.
Not entirely sure if this is the right move to make, because in some
ways it complicates things like write_flag and read_flag.

The reason I changed it is because I have some "psuedo" flags and regs
that don't have an index or number actually assigned to them. For example,
in the docs there is a Z (zero) flag but there is also a flag called NZ
(non-zero). Now, NZ doesn't actually exist in the hardware, it's just the
opposite of Z. So I used to represent the flags with their bit position
in the u8. But, that doesn't make much sense to NZ since it has the same
bit position as Z... except it's meaning is opposite.