Lots of refactoring done. The system is working.
The major thing I did was get rid of the mem::Memory
stuff because I was forced to use Arc<RefCell<>>. The
reason was Memory needed mut access to other structs like
the video and audio. Instead, we have a central System
struct that handles the video, audio, memory, memory mapping,
etc.